# Allow list fusion for Text and Text.Lazy unpack